+++ unique_toolkit/unique_toolkit/app/__init__.py
@@ -1,3 +1,5 @@
+from .find_env_file import EnvFileNotFoundError as EnvFileNotFoundError
+from .find_env_file import find_env_file as find_env_file
 from .init_logging import init_logging as init_logging
 from .init_sdk import get_endpoint_secret as get_endpoint_secret
 from .init_sdk import init_sdk as init_sdk